requirements.txt
Flask
Flask-Cors
requests
pymongo
python-dotenv
gunicorn
celery
redis
